Pulling all this logic together to try and keep it separate from the other worker clients
